requirements.txt
dash>=2.17.0
dash-bootstrap-components>=1.6.0
dash-chat>=0.2.0
anthropic>=0.39.0
diskcache>=5.6.0
python-dotenv>=1.0.0
